argparse_manpage: a new python module
Let's separate the Manpage generating logic (the new 'argparse_manpage'
module) from the part which helps the user to override the setuptools
commands (continues to be named 'build_manpages').

Fixes: #63
