+++ argparse_manpage/__init__.py
@@ -1,0 +1,5 @@
+"""
+argparse_manpage project
+"""
+
+__version__ = '3'
