+++ argparse_manpage/compat.py
