+++ argparse_manpage/tooling.py
