+++ build_manpages/__init__.py
@@ -2,6 +2,5 @@
 Export some useful methods in top-level.
 """
 
-__version__ = '3'
-
+from argparse_manpage import __version__
 from .build_manpages import build_manpages, get_build_py_cmd, get_install_cmd
