+++ setup.cfg
@@ -1,3 +1,3 @@
 [build_manpages]
 manpages =
-    man/argparse-manpage.1:object=ap:pyfile=build_manpages/cli.py
+    man/argparse-manpage.1:object=ap:pyfile=argparse_manpage/cli.py
